atri/requirements.txt
nb-cli
nonebot-adapter-onebot

requests
aiohttp
aiofiles
numpy
scipy
opencv-python
